src/payment_gateway/requirements.txt
fastapi==0.70.0
orjson==3.6.4
uvicorn==0.15.0
gunicorn==20.1.0
uvloop==0.16.0
pydantic~=1.8.2
backoff~=1.11.1

fastapi-sqlalchemy==0.2.1
psycopg2-binary==2.9.2
asyncpg==0.25.0
alembic==1.7.7
